Added new header component to all pages
Added some code to the header template to work out which menu item needs to be selected.

Added a new variable to each page to state which menu item should be selected.

Replaced the old HTML nav with the new template on each page and added the active_page variable to the pages